refactor: extract shared download logic to rule_loader, remove _rules.py
- Move sha256, download, and status check functions into rule_loader.py
  as the single source of truth
- Rewrite scripts/download_rules.py as a thin wrapper
- Remove commands/_rules.py (scripts/download_rules.py is the only entry point)
- Add release recovery steps to README
